implemented table sorting
